requirements.txt
torch
torchaudio
transformers
sounddevice
numpy
soundfile
librosa
noisereduce
webrtcvad
python-multipart
uvicorn
fastapi 
pyngrok
pyyaml